Draw() outputs rectangles
Modified ALL draw functions so that body.draw() outputs a list of rectangles and shape.draw outputs a single rectangle. These can be (in theory) used to update the screen selectively to improve frame rate, however it appears that this functionality is broken in pygame. perhaps pygame_sdl2 could work